:art: Refactor parts of add patient form into seperate components
